Remove Azure restrict-vm-publisher policy
For the sake of a demo org, we don't really need it; it's not one we can
change during a demo, nor one we really talk about usually.

However, the main reason for removing it is that it fails with machine
images that we create ourselves:

```
module.webserver.azurerm_linux_virtual_machine.example has source_image_reference.0.publisher with value null that is not in the allowed list: [RedHat, Canonical, MicrosoftWindowsServer]
```

While I would ideally like to fix this, pragmatically it's not worth the
time right now.